+++ NEWS.md
@@ -3,8 +3,10 @@
 ### update fonctions
 - model_bh_intra_location
     - deal with vec_env_with_no_controls empty
-- format_data_PPBstats.data_network : clean check at the beginning
-
+- format_data_PPBstats.data_network : 
+    - clean check at the beginning
+    - add long and lat for bipart network
+    
 # **PPBstats 0.23**
 ### update fonctions
 - format_data_PPBstats.R :